Restructures the assets directory
