+++ assets/frontend.js
@@ -1,1 +1,2 @@
-import './frontend-styles.scss';
+// Entry point for Webpack compilation.
+import './global-styles/frontend-styles.scss';
